feat(apps): align My Apps and App Catalog table columns
Drop the Revision column from My Apps, rename Name to App Name in both
tables, and add a Repository column and the app icon in front of the
name to the catalog table, matching My Apps. The org/repo label helper
moves to utils as repoLabel for shared use.
